Restructure download code layout
